% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{scExample}
\alias{scExample}
\title{Example expression matrix for scTHI.}
\format{An object of class \code{matrix} with 2000 rows and 100 columns.}
\usage{
scExample
}
\description{
A 2000 x 100 matrix from the wiki manual to showcase the use of scTHI
scExample
}
\keyword{datasets}
