% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI.R
\name{scTHI.plotResult}
\alias{scTHI.plotResult}
\title{scTHI.plotResults}
\usage{
scTHI.plotResult(scTHIresult, cexNames = 0.8, plotType = c("score",
  "pair"), legendPos = c("topleft", "topright"))
}
\description{
This function  ...
}
\examples{

scTHI.plotResult
}
