% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI.R
\docType{package}
\name{scTHI}
\alias{scTHI}
\alias{scTHI-package}
\title{single cell Tumor Hist Interaction (scTHI)}
\description{
Indentification of  significantly activated ligand-receptor 
interactions across clusters of cells from single-cell RNA sequencing data.
Single-cell RNA sequencing is the reference technique to characterize the heterogeneity of tumor microenvironment. 
The composition of the various cell types making up the microenvironment can significantly affect the way 
in which the immune system activates cancer rejection mechanisms. Understanding the cross-talk signals
 between immune cells and cancer cells is a fundamental for the identification immuno-oncology therapeutic targets. 
scTHI is  a novel method, single cell Tumor-Host Interaction tool (scTHI), 
to identify significantly activated ligand-receptor interactions across clusters of cells from 
single-cell RNA sequencing data. 
cTHI is based on the hypothesis that when patterns of interaction are active, 
they are also simultaneously and highly expressed in
homogeneous cell populations. We also model the autocrine and paracrine 
signalling effects of L-R partners
}
\details{
Please have a look at the vignette for a in-depth introduction to the package.
}
