% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI.R
\name{scTHI.runTsne}
\alias{scTHI.runTsne}
\title{scTHI.runTsne}
\usage{
scTHI.runTsne(scTHIresult)
}
\arguments{
\item{scTHIresult}{scTHI object.

scTHI.runTsne}
}
\description{
Runs t-SNE dimensionality reduction on selected features based on Rtsne package.
}
