% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI.R
\name{scTHI.score}
\alias{scTHI.score}
\title{scTHI.score}
\usage{
scTHI.score(expMat, cellCusterA, cellCusterB, cellCusterAName,
  cellCusterBName, topRank = 10, autocrineEffect = TRUE,
  fileNameBase = "scTHI", filterCutoff = 0.5, PValue = TRUE,
  pvalueCutoff = 0.05, nPermu = 1000, ncore = 8)
}
\arguments{
\item{expMat}{ScRNA-seq gene expression matrix where rows are genes presented with Hugo Symbols and columns are cells. Gene expression values should be counts or normalized counts.}

\item{cellCusterAName}{A character string labeling the clusterA.}

\item{cellCusterBName}{A character string labeling the clusterB.}

\item{topRank}{Filter threshold. Set to 10 (default) means that each gene of the interaction pair will be considered as expressed in a cell if it's in the top rank 10 percent.}

\item{fileNameBase}{Project name.}

\item{filterCutoff}{Score threshold (default is 0.50). For each interaction pair, if the score calculated (for the partnerA or partnerB) will be less than filterCutoff the interaction pair will be discarded.}

\item{PValue}{Logical, set to TRUE (default) compute statistical iterations. If p.value < 0.05, the value will be returned.}

\item{nPermu}{Number of iterations to perform (default is 1000).}

\item{ncore}{Number of processors to use.}

\item{clusterA}{Vector of columns of expMat that belong to the first cluster.}

\item{clusterB}{Vector of columns of expMat that belong to the second cluster.}

\item{foldChange}{Logical, default is TRUE. This argument allows to calculate a fold change for each gene of the interaction pair (PartenerA and PartnerB) with respect to a third cluster of cells, specified in cellsToCompare.}

\item{useData}{Logical, default is FALSE. Set to TRUE allows the user to compute fold change and pvalue later, providing as input a score matrix previously calculated.}

\item{dataToUse}{Score matrix to be provided as input if dataToUse is TRUE.}

\item{cellsToCompare}{Vector of columns of expMatToCompare that belong to the third cluster.}

\item{cellsToCompareNAME}{A character string labeling the third cluster.}

\item{expMatToCompare}{ScRNA-seq gene expression matrix of the third cluster cells. It may be the same as expMat.}
}
\description{
This function allows the user to compute a score for a set of ligand-receptor pairs, from a single cell gene expression matrix,
and detect specific Tumor-Host interactions. You must specify at least two clusters of cells (for example tumor cells and immune cells).
}
\examples{

scTHI
}
\keyword{interaction}
