% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI_runTsne.R
\name{scTHI_runTsne}
\alias{scTHI_runTsne}
\title{scTHI_runTsne}
\usage{
scTHI_runTsne(scTHIresult)
}
\arguments{
\item{scTHIresult}{scTHI object.}
}
\value{
The same object as scTHI_score with a fifth item tsneData
  (data.frame)
}
\description{
Runs t-SNE dimensionality reduction on selected features based on Rtsne
package.
}
\examples{
library(scTHI.data)
data(scExample)
result <-  scTHI_score(scExample,
                       cellCusterA = colnames(scExample)[1:30],
                       cellCusterB = colnames(scExample)[31:100],
                       cellCusterAName = "ClusterA",
                       cellCusterBName = "ClusterB", filterCutoff = 0,
                       pvalueCutoff = 1, nPermu = 100, ncore = 8)
result <- scTHI_runTsne(result)
}
