% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI.R
\name{TME.classification}
\alias{TME.classification}
\title{TME.classification}
\usage{
TME.classification(expMat, ncore = 48, minLenGeneSet = 10,
  pvalFilter = FALSE, fdrFilter = TRUE, pvalCutoff = 0.01,
  nesCutoff = 0.58, nNES = 1)
}
\description{
This function  ...
}
\examples{

TME.classification
}
