% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI.R
\name{TME.plot}
\alias{TME.plot}
\title{TME.plot}
\usage{
TME.plot(tsneData, Class, cexPoint = 0.8)
}
\arguments{
\item{tsneData}{X and y coordinates of points in the plot.}

\item{Class}{Object returned by TME.classification function.}

\item{cexPoint}{Set the point size.

TME.plot}
}
\description{
Generates a plot on the t-SNE coordinates, labeling cells by TME classification.
}
