% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI.R
\name{scTHI.plotCluster}
\alias{scTHI.plotCluster}
\title{scTHI.plotCluster}
\usage{
scTHI.plotCluster(scTHIresult, cexPoint = 0.8, legendPos = c("topleft",
  "topright", "bottomright", "bottomleft"))
}
\description{
This function  ...
}
\examples{

scTHI.plotCluster
}
