% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI.R
\name{scTHI}
\alias{scTHI}
\title{scTHI}
\usage{
scTHI(expMat, cellCusterA, cellCusterB, cellCusterAName, cellCusterBName,
  topRank = 10, fileNameBase = "CustomCellPhone", filterCutoff,
  foldChange = TRUE, useData = FALSE, dataToUse, cellsToCompare,
  cellsToCompareNAME, expMatToCompare, PValue = TRUE, nPermu = 1000,
  ncore = 8)
}
\arguments{
\item{expMat}{expression matrix.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
scTHI
}
\keyword{interaction}
