% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI.R
\name{scTHI}
\alias{scTHI}
\title{scTHI}
\usage{
scTHI(expMat, cellCusterA, cellCusterB, cellCusterAName, cellCusterBName,
  topRank = 10, fileNameBase = "CustomCellPhone", filterCutoff,
  foldChange = TRUE, useData = FALSE, dataToUse, cellsToCompare,
  cellsToCompareNAME, expMatToCompare, PValue = TRUE, nPermu = 1000,
  ncore = 8)
}
\arguments{
\item{expMat}{expression matrix.}

\item{cellCusterBName}{}

\item{clusterA}{vector of columns of expMat that belong to the first cluster}

\item{clusterB}{vector of columns of expMat that belong to the second cluster}
}
\description{
This function computes a set of ligand-receptor interactions from a single cell gene expression matrix.
you must specify at least two clusters of cells (for example tumor cells and immune cells)
}
\examples{

scTHI
}
\keyword{interaction}
