% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTHI.R
\name{TME.plot}
\alias{TME.plot}
\title{TME.plot}
\usage{
TME.plot(tsneData, Class, cexPoint = 0.8)
}
\arguments{
\item{tsneData}{X and y coordinates of points in the plot.}

\item{Class}{Object returned by TME.classification function.}

\item{cexPoint}{Set the point size.}
}
\value{
None
}
\description{
Generates a plot on the t-SNE coordinates, labeling cells by TME
classification.
}
\examples{
data(scExample)
result <-  scTHI.score(scExample, cellCusterA = colnames(scExample)[1:30],
                    cellCusterB = colnames(scExample)[31:100],
                    cellCusterAName = "ClusterA",
                    cellCusterBName = "ClusterB", filterCutoff = 0,
                    pvalueCutoff = 1, nPermu = 100, ncore = 8)
result <- scTHI.runTsne(result)
Class <- TME.classification(scExample)
TME.plot(tsneData = result$tsneData, Class)
}
